+++ 2021 Progetto Sistemi Operativi/Istruzioni progetti 2021.md
@@ -1,0 +1,9 @@
+# Progetto Laurea in Informatica UniPR esame Febbraio 2021 Sistemi Operativi
+
+Progetto svolto da Andrea Ciccarello
+
+### Per testare con le 3 matrici di esempio 
+
+1. Prendere le matrici A e B con la dimensione voluta tra quelle proposte e rinominarle in MatriceA.txt e MatriceB.txt
+2. Bisogna modificare la variabile di preprocesing DimensioneStandard nella dimensione voluta per esempio se si vuole provare con una matrice 10x10 allora DimensioneStandard 10
+
