+++ 2021 Progetto Sistemi Operativi/Matrici-di-prova/matriceA_3.txt
@@ -1,0 +1,3 @@
+83 21 82 
+55 36 18 
+39 9 1
