+++ 2021 Progetto Sistemi Operativi/Matrici-di-prova/matriceB_100.txt
@@ -1,0 +1,20 @@
+9 9 85 23 53 26 84 2 4 54 33 13 25 16 26 14 67 30 40 62 
+30 9 84 76 92 25 54 89 75 32 36 28 31 75 5 78 55 64 24 81 
+2 35 14 28 34 68 29 78 10 60 83 15 67 10 94 93 39 18 21 54 
+36 59 84 0 95 14 31 80 96 13 79 30 60 68 42 86 26 32 85 52 
+76 20 59 64 38 59 68 26 13 79 2 88 39 5 22 22 39 93 7 88 
+35 53 66 26 46 34 45 16 21 27 98 23 54 34 86 40 45 6 3 21 
+67 70 32 94 85 35 70 39 38 95 7 71 55 46 75 15 66 11 14 32 
+87 85 8 77 66 12 2 51 92 99 26 14 90 14 11 49 93 36 54 60 
+9 28 97 56 30 99 46 75 63 50 62 22 72 19 6 94 50 11 15 64 
+2 76 35 18 61 20 42 59 42 27 46 32 67 32 90 88 19 31 94 4 
+33 11 3 74 50 71 50 96 70 36 21 69 14 66 28 14 28 96 30 94 
+29 20 37 90 70 19 56 13 64 80 16 54 8 54 31 59 44 10 27 25 
+44 78 11 64 44 9 80 96 24 65 33 75 18 48 93 74 39 10 65 23 
+5 25 66 1 22 69 94 35 9 91 10 4 43 64 79 27 4 85 59 29 
+49 89 75 38 61 76 51 44 24 39 11 7 0 33 70 53 19 11 3 67 
+52 17 79 22 13 63 55 41 8 82 78 21 22 16 31 70 90 27 90 96 
+83 44 24 44 11 7 2 98 52 79 78 67 59 14 64 24 98 91 92 32 
+36 75 68 16 7 16 88 60 46 51 53 93 55 68 31 64 2 95 78 93 
+24 23 42 76 73 92 97 13 56 80 62 10 61 44 84 78 0 97 47 68 
+99 77 67 28 78 4 81 29 78 62 46 0 3 17 12 0 78 44 60 68
