+++ 2021 Progetto Sistemi Operativi/Matrici-di-prova/matriceB_3.txt
@@ -1,0 +1,3 @@
+32 17 58 
+20 5 28 
+47 20 27
